First pass at windows support
